+++ lab06/structures/data.go
@@ -4,9 +4,9 @@ import "time"
 
 type Data struct {
 	Date   time.Time
-	Last   float32
+	Last   float64
 	Volume int32
-	Open   float32
-	High   float32
-	Low    float32
+	Open   float64
+	High   float64
+	Low    float64
 }
